Camera Rotation & Refactor
Made the camera rotate when holding Alt, when you release alt the camera snaps back to its default position. Camera's movement is now based on this rotation, we do this by rotating the gimbal only on the x axis and the root object on the y axis, then moving based on the root's transform.

Started on a pan but it's not great, have a better pan in mind but in the mean time I started work on a new infantry squad; Marines. Refactored unit files around to organize infantry and vehicles into their own folders and split the infantry code into roughnecks and base infantry code, reducing infantry member from 700 lines to 300 and roughneck to 300, eliminating stench.

Got Roughnecks working with their own objects now, haven't started on any Marine code yet.
