+++ Assets/Materials/Units/RoughneckMat.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 63d1a981ce93d254790e90f24592572c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
