+++ Assets/Units/Infantry.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6d7b966867524934dbf09532e2b8ffdc
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
