+++ Assets/Units/Infantry/InfantryMember.cs.meta
