+++ Assets/Units/Infantry/Marine.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5a51a63e83d60534789298b02ec77dcd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
