+++ Assets/Units/Infantry/Roughneck.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f898798f90ea9984ab67b93e79f23bcf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
