+++ Assets/Units/Infantry/RoughneckSquad.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bd88b885457ea1b478b55fb5275fe5cd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
